cleanup contains_keyword_match
